+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [0.0.11] - 2025-04-07
+
 ### 新增
 
 - MyBatis
@@ -106,7 +108,8 @@
 - Issue 管理
   - 🔥添加Issue行标记，并可在浏览器中快速打开，支持Git Issue和Jira。(#10)
 
-[Unreleased]: https://github.com/iimik/final-aio/compare/v0.0.10...HEAD
+[Unreleased]: https://github.com/iimik/final-aio/compare/v0.0.11...HEAD
+[0.0.11]: https://github.com/iimik/final-aio/compare/v0.0.10...v0.0.11
 [0.0.10]: https://github.com/iimik/final-aio/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/iimik/final-aio/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/iimik/final-aio/compare/v0.0.7...v0.0.8